% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePhysioMap.R
\name{calculatePhysioMap}
\alias{calculatePhysioMap}
\title{Mapping new data into a physiological-space}
\usage{
calculatePhysioMap(InputData, Space, GenesRatio = 0.05, PARALLEL = FALSE,
  TTEST = FALSE, WilxSTATICResponse = F, NumbrOfCores = NA)
}
\arguments{
\item{InputData}{A matrix with genes as rows and samples as columns. Corresponding Entrez Gene
IDs must be assigned to 'rownames' of the matrix, and name of each sample/column should be written in 'colnames'.}
}
\description{
calculatePhysioMap computes mapped values of each input sample inside of a space,
calculated prior from a compendium of known samples (Check ?SpaceMaker for more info).
}
\examples{
print()

}
