% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePhysioMapCore_foreach.R
\name{calculatePhysioMapCore_foreach}
\alias{calculatePhysioMapCore_foreach}
\title{Mapping new data into a physiological-space using foreach}
\usage{
calculatePhysioMapCore_foreach(InputData, Space, NSamples, GenesRatio, NGenes,
  STATICResponse, pb, TTEST)
}
\description{
calculatePhysioMapCore_foreach is an internal function of calculatePhysioMap,
computing the main mapping. It's not recommended to use calculatePhysioMapCore_foreach by the user outside of
calculatePhysioMap.
}
