% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePhysioMapCore_parCapply.R
\name{calculatePhysioMapCore_parCapply}
\alias{calculatePhysioMapCore_parCapply}
\title{Mapping new data into a physiological-space using parCapply of parallel package}
\usage{
calculatePhysioMapCore_parCapply(InputData, Space, NSamples, GenesRatio, NGenes,
  STATICResponse, pb, TTEST, cl)
}
\description{
calculatePhysioMapCore_parCapply is an internal function of calculatePhysioMap,
computing the main mapping. It's not recommended to use calculatePhysioMapCore_parCapply by the user outside of
calculatePhysioMap.
}
