% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePhysioMapCores.R
\name{calculatePhysioMapCores}
\alias{calculatePhysioMapCores}
\title{Dispatching the parallel calculation among core functions}
\usage{
calculatePhysioMapCores(InputData, Space, NSamples, GenesRatio, NGenes,
  STATICResponse, pb, TTEST, cl, ParallelMethod = "parCapply")
}
\arguments{
\item{InputData}{A matrix, SummarizedExperiment object or a list,
based on the gene expression data (or any other type of high
dimnesional, e.g. protein abundance, SNP, Methylation, etc.),
to be analysed. InputData has to have a specific format to be
properly analysed, these requirements are thoroughly explained
in the 'Details' section.}

\item{Space}{The space in which the 'InputData' will be mapped. Just as
'InputData', it should be a matrix with genes as rows and samples as
columns, with corresponding Entrez Gene IDs in 'rownames' of the matrix,
and name of each axis of the space written in 'colnames'.}

\item{NSamples}{Number of samples (columns) in InputData.}

\item{GenesRatio}{The ratio of gene expression values to be considered in
the calculation. In high dimensional omics data, signal to noise ratio
has a direct relation with the relative magnitude of expressions.
We aim to remove the noisy genes, hence we only keep the "GenesRatio*100"
percent highest and lowest gene expression values of each sample.
GenesRatio should be a numerical value between 0 and 1. Default value
is 0.05.}

\item{NGenes}{Number of genes (rows) in InputData or Space.}

\item{STATICResponse}{Logical value indicating if 'statistic' should be
returned rather than the default 'signed p value'. Default value is FALSE.}

\item{pb}{Progress bar made by progress::progress_bar$new.}

\item{TTEST}{Logical value indicating if t.test should be done in place
of the default wilcoxon rank-sum test (more info can be found
in the original PhysioSpace: Lenz et. al., PLOS One 2013). Using t.test
will speed up calculations. Default value is FALSE.}

\item{cl}{Cluster made by parallel::makeCluster.}

\item{ParallelMethod}{Parallel method to use.
Two methods are implemented so far: "parCapply"
which uses parCapply function of parallel package, and "foreach" which
uses foreach package to process in parallel.
Speed-wise, foreach has an edge on parCapply, but the latter
is more stable. Hence, we recommend to use parCapply.
The default value for ParallelMethod is "parCapply".}
}
\value{
Matrix of mapped 'InputData' values in 'Space', with
rows corresponding to axes of 'Space' and columns representing
samples in 'InputData'.
}
\description{
calculatePhysioMapCores is an internal
function of calculatePhysioMap, that calls the proper internal
function for computing the main mapping in parallel. We don't
recommend the use of calculatePhysioMapCores outside of
calculatePhysioMap.
}
