% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InptChecker.R
\name{.inptChecker}
\alias{.inptChecker}
\title{Checking calculatePhysioMap Inputs}
\usage{
.inptChecker(InputData, Space)
}
\arguments{
\item{InputData}{A matrix, SummarizedExperiment object or a list,
based on the gene expression data (or any other type of high
dimensional data, e.g. protein abundance, SNP, Methylation, etc.),
to be analysed. InputData has to have a specific format to be
properly analysed, these requirements are thoroughly explained
in the 'Details' section of calculatePhysioMap() function.}

\item{Space}{The space in which the 'InputData' will be mapped. Just as
'InputData', it should be a matrix with genes as rows and samples as
columns, with corresponding Entrez Gene IDs in 'rownames' of the matrix,
and name of each axis of the space written in 'colnames'.}
}
\value{
.inptChecker returns corrected 'InputData' and 'Space'
directly to the environment it was called from
(By assigning new matrices to parent.frame()).
}
\description{
.inptChecker is an internal function used by
calculatePhysioMap for checking the format of inputs, 'InputData'
and 'Space' to be exact. It 1- checks to see if both 'InputData'
and 'Space' are matrices, and 2- matches the rows of 'InputData'
or 'Space' based on their row names.
}
\examples{
 SimulatedGeneExpressionData <-
   matrix(
     rnorm(n = 100000, mean = 0, sd = 100),
     ncol = 10,
     dimnames = list(1:10000, 1:10)
   )
 PhysioSpaceMethods:::.inptChecker(InputData = SimulatedGeneExpressionData[, 1:5],
              Space = SimulatedGeneExpressionData[sample(1:10000), 6:10])

}
