% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InptPreparer.R
\name{.inptPreparer}
\alias{.inptPreparer}
\title{Preparing InputData for Calculation}
\usage{
.inptPreparer(InputData)
}
\arguments{
\item{InputData}{A matrix, SummarizedExperiment object or a list,
based on the gene expression data (or any other type of high
dimnesional, e.g. protein abundance, SNP, Methylation, etc.),
to be analysed. InputData has to have a specific format to be
properly analysed, these requirements are thoroughly explained
in the 'Details' section.}
}
\value{
.inptPreparer returns InputData as a matrix
with proper row- and column names.
}
\description{
.inptPreparer is an internal function used by
.inptChecker and spaceMaker to set up row- and column names of
'InputData', and convert it to a matrix, if necessary.
}
\examples{
 SimulatedGeneExpressionData <- matrix(
   rnorm(n = 100000, mean = 0,
         sd = 100),
   ncol = 10,
   dimnames = list(1:10000, 1:10)
 )
 SimulatedGeneExpressionData_checked <-
     PhysioSpaceMethods:::.inptPreparer(SimulatedGeneExpressionData)

 library(SummarizedExperiment)
 SimulatedGeneExpressionData_SE <- SummarizedExperiment(
     assays = list(GEX = SimulatedGeneExpressionData),
     rowData = data.frame("EntrezID" =
                              rownames(SimulatedGeneExpressionData)),
     colData = data.frame("SampleName" =
                              colnames(SimulatedGeneExpressionData))
 )
 SimulatedGeneExpressionData_SE_checked <-
     PhysioSpaceMethods:::.inptPreparer(SimulatedGeneExpressionData_SE)

}
