% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelInitializer.R
\name{.parallelInitializerSolo}
\alias{.parallelInitializerSolo}
\title{Initializing and Preparing for Parallel Processing on a Single Machine}
\usage{
.parallelInitializerSolo(NumbrOfCores)
}
\arguments{
\item{NumbrOfCores}{Number of cores (Threads) to use in parallel.}
}
\value{
An object of class c("SOCKcluster", "cluster"), made by
makeCluster of package parallel.
}
\description{
.parallelInitializerSolo is an internal
function of .parallelInitializer.
}
\examples{
PhysioSpaceMethods:::.parallelInitializerSolo(NumbrOfCores = 2)

}
