% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InptChecker.R
\name{inptChecker}
\alias{inptChecker}
\title{Checking calculatePhysioMap Inputs}
\usage{
inptChecker(InputData, Space, ImputationMethod = "PCA")
}
\arguments{
\item{InputData}{A matrix, SummarizedExperiment object or a list,
based on the gene expression data (or any other type of high
dimnesional, e.g. protein abundance, SNP, Methylation, etc.),
to be analysed. InputData has to have a specific format to be
properly analysed, these requirements are thoroughly explained
in the 'Details' section.}

\item{Space}{The space in which the 'InputData' will be mapped. Just as
'InputData', it should be a matrix with genes as rows and samples as
columns, with corresponding Entrez Gene IDs in 'rownames' of the matrix,
and name of each axis of the space written in 'colnames'.}

\item{ImputationMethod}{Imputation method to use in case of missing
values. Available methods are "PCA" and "KNN". Default is "PCA".}
}
\value{
inptChecker returns corrected 'InputData' and 'Space'
directly to the environment it was called from
(By assigning new matrices to parent.frame()).
}
\description{
inptChecker is an internal function used by
calculatePhysioMap to check the format of inputs, 'InputData'
and 'Space' to be exact. It checks to see 1- both 'InputData'
and 'Space' are matrices, 2- if 'InputData' or 'Space' contain
any NA, and trying to impute using imputeMissingGeneExpression
function if they do, and 3- it matches the rows of 'InputData'
or 'Space' based on their row names.
}
\examples{
 SimulatedGeneExpressionData <-
   matrix(
     rnorm(n = 100000, mean = 0, sd = 100),
     ncol = 10,
     dimnames = list(1:10000, 1:10)
   )
 inptChecker(InputData = SimulatedGeneExpressionData[, 1:5],
              Space = SimulatedGeneExpressionData[sample(1:10000), 6:10])

}
