% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelInitializer.R
\name{parallelInitializer}
\alias{parallelInitializer}
\title{Creates and Uses Physio Spaces as a dimension reduction mapping}
\usage{
parallelInitializer(NumbrOfCores)
}
\arguments{
\item{CLUSTER, machineAddresses, rscript, NumbrOfCores}{}
}
\description{
This package uses 'Big Data' to make robust 'Physiological Vectors' in N dimensions spaces, with which you will map new data to extract infromation from a big high dimensional confusing new data.
}
\examples{
parallelInitializer()

}
