% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelInitializer.R
\name{parallelInitializer.SingleMachine}
\alias{parallelInitializer.SingleMachine}
\title{Initializing and Preparing for Parallel Processing on a Single Machine}
\usage{
parallelInitializer.SingleMachine(NumbrOfCores)
}
\arguments{
\item{NumbrOfCores}{Number of cores (Threads) to use in parallel.}
}
\value{
An object of class c("SOCKcluster", "cluster"), made by makeCluster of package parallel.
}
\description{
parallelInitializer.SingleMachine is an internal function of parallelInitializer.
}
\examples{
parallelInitializer.SingleMachine(NumbrOfCores = 2)

}
