% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMapPlotter.R
\name{PhysioHeatmap}
\alias{PhysioHeatmap}
\title{Drawing Heatmap of CalculatePhysioMap()'s Output}
\usage{
PhysioHeatmap(PhysioResults, ColorLevels = 100, PDFFullName, PDFWidth = 7,
  PDFHeight = 7, main = "")
}
\arguments{
\item{PhysioResults}{Matrix of scores generated by CalculatePhysioMap().}

\item{ColorLevels}{An integer indicating how many colors to use when plotting the heatmap. Default is 100.}

\item{PDFFullName}{Full name (including the path) of the PDF file in which the heatmap is saved, e.g. "~/Desktop/outheat.pdf". For now
only exporting to a PDF is supported, so this parameter needs to be set or else the function breaks.}

\item{PDFWidth}{Width of the output PDF, in inches. Default is 7.}

\item{PDFHeight}{Height of the output PDF, in inches. Default is 7.}

\item{main}{The title of the heatmap. Default is an empty string (no title).}
}
\value{
PhysioHeatmap returns(Invisibly) PDFFullName.
}
\description{
Draws a custom heatmap based on the result matrix generated by CalculatePhysioMap() function.
}
\examples{
require(PhysioSpacesMethods)
randMatInpt <- matrix(data = rnorm(n = 4000,mean = 10,sd = 20),nrow = 400)
rownames(randMatInpt) <- paste("ROWS",1:400)
colnames(randMatInpt) <- paste("Sample",1:10)

randMatRef <- matrix(data = rnorm(n = 12000,mean = 10,sd = 20),nrow = 400)
rownames(randMatRef) <- paste("ROWS",1:400)
colnames(randMatRef) <- paste("Space",1:30)

res <- calculatePhysioMap(InputData = randMatInpt, Space = randMatRef)

PhysioHeatmap(PhysioResults = res, PDFFullName = "~/Desktop/outheat.pdf", main = "Heatmap Testing")
}
