% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTestWrapper.R
\name{tTestWrpr}
\alias{tTestWrpr}
\title{T-testing Between Plus and Minus Genes}
\usage{
tTestWrpr(ReferencesJ, iplus, iminus, STATICResponse)
}
\arguments{
\item{ReferencesJ}{Vector of gene expressions to do statistical test on.}

\item{iplus}{Index of first group of genes for statistical testing.}

\item{iminus}{Index of second group of genes for statistical testing.}

\item{STATICResponse}{Same STATICResponse as in calculatePhysioMap.
Check calculatePhysioMap's help for more info.}
}
\value{
Log2 signed p value of t-test if STATICResponse==FALSE,
t-test statisitic if STATICResponse==TRUE.
}
\description{
tTestWrpr is an internal function used by
calculatePhysioMap. It does a t-test between iplus and iminus
indexed genes in ReferencesJ.
}
\examples{
 SimulatedReferenceSpace <-
   matrix(
     rnorm(n = 100000, mean = 0, sd = 100),
     ncol = 10,
     dimnames = list(1:10000, 11:20)
   )
 tTestWrpr(
   ReferencesJ = SimulatedReferenceSpace[, 4],
   iplus = sample(
     1:nrow(SimulatedReferenceSpace),
     size = nrow(SimulatedReferenceSpace) / 20
   ),
   iminus = sample(
     1:nrow(SimulatedReferenceSpace),
     size = nrow(SimulatedReferenceSpace) / 20
   ),
   STATICResponse = FALSE
 )

}
