% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WilcoxonTestWrapper.R
\name{wilTestWrapper}
\alias{wilTestWrapper}
\title{Creates and Uses Physio Spaces as a dimension reduction mapping}
\usage{
wilTestWrapper(ReferencesJ, iplus, iminus, WilxSTATICResponse)
}
\arguments{
\item{ReferencesJ, iplus, iminus, WilxSTATICResponse}{}
}
\description{
This package uses 'Big Data' to make robust 'Physiological Vectors' in N dimensions spaces, with which you will map new data to extract infromation from a big high dimensional confusing new data.
}
\examples{
wilTestWrapper()

}
