% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePhysioMap.R
\name{calculatePhysioMap}
\alias{calculatePhysioMap}
\title{Mapping new data into a physiological-space}
\usage{
calculatePhysioMap(InputData, Space, GenesRatio = 0.05,
  PARALLEL = FALSE, NumbrOfCores = NA, TTEST = FALSE,
  STATICResponse = FALSE, ImputationMethod = "PCA",
  ParallelMethod = "parCapply")
}
\arguments{
\item{InputData}{A matrix of input gene expressions to be analyzed, with genes as rows and samples as columns. Corresponding Entrez Gene
IDs must be assigned to 'rownames' of the matrix, and name of each sample/column should be written in 'colnames'. REMEMBER that the gene
expressions in 'InputData' should be relative; e.g. fold change or signed p value of a statistical test.}

\item{Space}{The space in which the 'InputData' will be mapped. Just as 'InputData',
it should be a matrix with genes as rows and samples as columns, with corresponding Entrez Gene
IDs in 'rownames' of the matrix, and name of each axis of the space written in 'colnames'.}

\item{GenesRatio}{The ratio of gene expression values to be considered in the calculation. In the calculations only the highest and lowest
GenesRatio*100 percent of values in each sample are used, since signal to noise ratio in gene expression values
has a direct relation to the relative magnitute of expressions and out noisy inputs need to be filtered out.
GenesRatio should be a numerical value between 0 and 1. Default value is 0.05.}

\item{PARALLEL}{Logical value indicating if calculations should be done in parallel. Default value is FALSE.
It is not recomemnded to use PARALLEL=TRUE on small datasets since due to large overhead, it could take more time than
using PARALLEL=FALSE.}

\item{NumbrOfCores}{Number of cores to be used when 'PARALLEL' is TRUE. Default is NA which will result in the program using
'all available cores - 1'. Assigning a number higher that the value parallel::detectCores() returns will result in an error.}

\item{TTEST}{Logical value indicating if t.test should be done in place of the default wilcoxon rank-sum test (more info can be found
in the original PhysioSpace: Lenz et. al., PLOS One 2013). Using t.test will speed up calculations. Default value is FALSE.}

\item{STATICResponse}{Logical value indicating if 'statistic' should be returned rather than the default 'signed p value'.
Default value is FALSE.}

\item{ImputationMethod}{Imputation method to use in case of missing values.}

\item{ParallelMethod}{Parallel method to use when PARALLEL=TRUE. Two methods are implemented so far: "parCapply"
which uses parCapply function of parallel package, and "foreach" which uses foreach package to process in parallel.
Speed-wise, foreach has an edge on parCapply, but the latter is more stable. Hence we recommend parCapply.
The default value for ParallelMethod is "parCapply".}
}
\value{
Matrix of mapped 'InputData' values in 'Space', with rows corrisponding to axises of 'Space' and columns representing
samples in 'InputData'. Mapped values are signed p value when STATICResponse==FALSE, and are 'statistic' value when
STATICResponse==TRUE (more info can be found in the original PhysioSpace: Lenz et. al., PLOS One 2013).
}
\description{
calculatePhysioMap computes mapped values of each input sample inside of a space,
calculated prior from a compendium of known samples.
}
\details{
PhysioSpace is a robust
statistical method for relating high dimensional omics data sets from heterogeneous sources using shared physiological
processes. It is designed to take advantage of the vast availability of public omics data, which in combination with
statistical approaches makes a potent tool capable of analyzing heterogeneous biological data sets.
'calculatePhysioMap' is the main analytical function of the package. It uses a nonlinear mapping function to relate the
unknown input data with an physiological space. Physiological spaces are mathematical spaces build upon known
physiological data, and built using the 'spaceMaker' function.
}
\examples{
 SimulatedGeneExpressionData <- matrix(
   rnorm(n = 100000, mean = 0,
         sd = 100),
   ncol = 10,
   dimnames = list(1:10000, 1:10)
 )
 SimulatedReferenceSpace <- matrix(
   rnorm(n = 100000, mean = 0,
         sd = 100),
   ncol = 10,
   dimnames = list(1:10000, 11:20)
 )
 calculatePhysioMap(InputData = SimulatedGeneExpressionData,
                    Space = SimulatedReferenceSpace)
 if (parallel::detectCores() > 1) {
   #More than one core is needed for parallel processing
   calculatePhysioMap(
     InputData = SimulatedGeneExpressionData,
     Space = SimulatedReferenceSpace,
     PARALLEL = TRUE,
     NumbrOfCores = 2,
     GenesRatio = 0.01,
     STATICResponse = FALSE,
     TTEST = TRUE
   )
 }

}
\references{
Lenz, M., Schuldt, B. M., Müller, F. J., & Schuppert, A. (2013). PhysioSpace: relating gene expression
experiments from heterogeneous sources using shared physiological processes. PLoS One, 8(10), e77627.
}
