% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeMissingGeneExpression.R
\name{imputeMissingGeneExpression}
\alias{imputeMissingGeneExpression}
\title{Imputing Missing Data}
\usage{
imputeMissingGeneExpression(InptGEX, METHOD = "PCA")
}
\arguments{
\item{InptGEX}{Input matrix with missng values.}

\item{METHOD}{Method to use in imputation. Available methods are
KNN and PCA. Default is 'PCA'.}
}
\value{
A matrix with the same dimensions as InptGEX, with missing
values imputed.
}
\description{
imputeMissingGeneExpression is an internal function called
by inputChecker. It uses different methods to impute any missing value
of the InputData or Space (or else missing values would break the
pipeline in calculatePhysioMap).
}
\examples{
 MatToImpute <-
   matrix(
     rnorm(n = 100000, mean = 0, sd = 100),
     ncol = 10,
     dimnames = list(1:10000, 1:10)
   )
 MatToImpute[sample(x = 1:length(MatToImpute),
                    size = length(MatToImpute) / 20)] <- NA
 ImputedMat <-
   imputeMissingGeneExpression(InptGEX = MatToImpute, METHOD = "PCA")
 ImputedMat2 <-
   imputeMissingGeneExpression(InptGEX = MatToImpute, METHOD = "KNN")

}
