% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputChecker.R
\name{inputChecker}
\alias{inputChecker}
\title{Checking calculatePhysioMap Inputs}
\usage{
inputChecker(InputData, Space, ImputationMethod)
}
\arguments{
\item{InputData}{Same InputData as in calculatePhysioMap. Check calculatePhysioMap's help for more info.}

\item{Space}{Same Space as in calculatePhysioMap. Check calculatePhysioMap's help for more info.}

\item{ImputationMethod}{Imputation method to use in case of missing values.}
}
\value{
inputChecker returns corrected 'InputData' and 'Space' directly to the environment it was called from
(By assigning new matrices to parent.frame()).
}
\description{
inputChecker is an internal function used by calculatePhysioMap to ckeck the format of inputs, 'InputData' and 'Space' to be
exact. It checks to see 1- both 'InputData' and 'Space' are matrices, 2- if 'InputData' or 'Space' contain any NA, and trying to impute
using imputeMissingGeneExpression function if they do, and 3- it matches the rows of 'InputData' or 'Space' based on their row names.
}
\examples{
SimulatedGeneExpressionData <- matrix(rnorm(n = 100000, mean = 0, sd = 100),ncol = 10, dimnames = list(1:10000,1:10))
inputChecker(InputData = SimulatedGeneExpressionData[,1:5], Space = SimulatedGeneExpressionData[sample(1:10000),6:10])

}
