% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleThreadOfPhysioCalc.R
\name{singleThreadOfPhysioCalc}
\alias{singleThreadOfPhysioCalc}
\title{Mapping one sample into a physiological-space}
\usage{
singleThreadOfPhysioCalc(INPT, Space, NSamples, GenesRatio, NGenes,
  STATICResponse, pb, TTEST)
}
\description{
SingleThreadOfPhysioCalc is an internal function of calculatePhysioMap,
computing the main mapping. It's not recommended to use SingleThreadOfPhysioCalc by the user outside of
calculatePhysioMap.
}
