% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceMaker.R
\name{spaceMaker}
\alias{spaceMaker}
\title{Creates PhysioSpaces}
\usage{
spaceMaker(inputMatrix, annotationFrame, DESIGN, CONTRAST, RETURNMODEL = F,
  MODEL = NULL, RETURNFD = F)
}
\arguments{
\item{InputData}{}
}
\description{
This package uses 'Big Data' to make robust 'Physiological Vectors' in N dimensions spaces, with which you will map new data to extract infromation from a big high dimensional confusing new data.
}
\examples{
 spaceMaker()

}
